Moved CLI code/functions to sericom-core
